--- iOS-FakeNFT-Extended/UI/Statistics/UserCollectionView.swift
@@ -1,54 +1,0 @@
-//
-//  UserCollectionView.swift
-//  iOS-FakeNFT-Extended
-//
-//  Created by Владислав on 14.04.2026.
-//
-import SwiftUI
-import Foundation
-
-struct UserCollectionView: View {
-    
-    let title: String
-    let count: Int
-    
-    private let columns = [
-        GridItem(.flexible(), spacing: 12),
-        GridItem(.flexible(), spacing: 12),
-        GridItem(.flexible(), spacing: 12)
-    ]
-    
-    var body: some View {
-        ScrollView {
-            LazyVGrid(columns: columns, spacing: 12) {
-                ForEach(0..<count, id: \.self) { _ in
-                    VStack(alignment: .leading, spacing: 6) {
-                        RoundedRectangle(cornerRadius: 12)
-                            .frame(height: 92)
-                            .foregroundStyle(.gray.opacity(0.25))
-                        
-                        Text("NFT CARD")
-                            .font(.subheadline)
-                            .bold()
-                        
-                        Text("1.78 ETH")
-                            .font(.caption)
-                            .foregroundStyle(.secondary)
-                    }
-                }
-            }
-            .padding()
-        }
-        .navigationTitle(title)
-        .navigationBarTitleDisplayMode(.inline)
-    }
-}
-
-#Preview {
-    NavigationStack {
-        UserCollectionView(
-            title: "Коллекция NFT",
-            count: 112
-        )
-    }
-}
